Criando Contatos no Admin

+++ components/admin/Button.js
@@ -31,15 +31,19 @@ export default function Button(props){
                 Router.push(`/admin/${props.model}/${props.id}`)
                 break
             case "delete":
-                await axios.delete(`${serverUrl}/admin/${props.model}/${props.id}`).then(res=>{
-                    alert(`Sucesso! ${props.model} com id: ${props.id} Deletado.`  )
-                    Router.push(`/admin/${props.model}`)
-                }).catch(err=>{alert("Deu ruim")})
+                if (confirm("Tem certeza que deseja excluir?")){
+                    await axios.delete(`${serverUrl}/admin/${props.model}/${props.id}`).then(res=>{
+                        alert(`Sucesso! ${props.model} com id: ${props.id} Deletado.`  )
+                        Router.push(`/admin/${props.model}`)
+                    }).catch(err=>{alert("Deu ruim")}) }
+                break
+            case "Teste":
+                console.log(`Teste OK ${props.id}`)
                 break
         
         }
     }
     return(
-        <button className={styles.button} onClick={handleClick} >{props.text}</button>
+        <button className={styles.button} onClick={ props.click ? props.click :handleClick } >{props.text}</button>
     )
 }
